feature: add multi-stop gradients
Introduce support for multi-stop gradients, by synthesizing a 1x256 image for a gradient with more than two stops and then using that image when rendering the path. The fragment shader has a new mode which samples from the image texture when rendering a gradient.

The public API is extended with paint constructors that take multiple stops. The existing interface and behavior is maintained, so this change should not cause any changes in existing software.

Multi-stop gradients are kept in the image store and owned by the Canvas. This means that new backends can support multi-stop gradients just by implementing the fragment shader mode.

There's a new example which shows the multistop gradients in action. Hopefully we can add pixel tests and convert this into a pixel test.