Fix polygon area calculation
The previous algorithm was incorrect because it didn't consider
the triangle of the last two points and the first point, or of the last
point and the first two points. It's been replaced with one that works
without triangle area at all:
https://stackoverflow.com/a/1165943

This fixes some shapes being "randomly" thickened by the AA algorithm.