Fix leaking FBOs with the OpenGL renderer
When rendering into an Image, we store the associated FBO (and depth
stencil RBO) in a hash in the GL renderer, indexed by the image id.
When deleting an image on the canvas, we should also delete the FBO/RBO,
instead of leaking it until the renderer is destroyed.

With this change it's even more important to not delete images that are
used as rendering targets until after flush().