Add support for applying a gaussian blur filter to images
This adds the

   pub fn filter_image(&mut self, target_image: ImageId, filter: ImageFilter, source_image: ImageId)

function to the renderer, where the ImageFilter is an enum that starts
with a gaussian blur and the standard deviation parameter.

The GL renderer renders this in two passes, a horizontal and a vertical
blur.

Apart from adding filter code path, this patch makes one other change to
the rendering interface: The drawing commands are now moved from the
canvas to the renderer. This is needed because then the second pass in
the blur filter can just change the Command's image parameter, as it
consumes it. Otherwise Command would have to implement Clone, which
would likely be expensive due to the Vec fields within.