Make TextContext public API
This makes the existing text API on the Canvas public in the scope of
`femtovg::TextContent`. This allows measuring text and line breaking
without any canvas, connection to the display server or OpenGL context.

This also makes it possible to share font data across Canvas instances
to save memory. Only the glyph textures aren't shared, but that's okay
for now.

For text measurements without a canvas it's important to note that any
scale (dpi or transform) that's usually applied on the canvas needs to
be applied manually. That however is only needed when for example the
dpi is != 1. On our specific use-case we do all device pixel ratio ahead
of femtovg.