Fix clearing of stencil buffer
When filling anti-aliased paths we draw the quad of the bounding box
first with the stencil test to achieve the desired anti-aliasing, and
afterwards we draw the quad a second time with zero as stencil op (and
mask according, etc.) to clear it. This clearing is likely faster than
clearing the entire stencil buffer, but if the boundaries are not
entirely correct, non-zero values might be left around the edges of the
shape in the stencil buffer.

When another shape is drawn over a shape that "leaks" into the stencil
buffer and that shape has an overall bigger bounding box, the fill of
that shape will not only be visible inside its actual shape but also in
the places where non-zero values were left over.

This patch attempts to correct that by adjusting the bounding box of the
path for the purpose of the quad fill by the fringe for anti-aliasing
that's not otherwise taken into account.