Fix rendering into FBOs with OpenGL ES 2.0
The stencil buffer attachment code on FBOs does not work with a pure GL
ES 2.0 context. The workaround of commit
141b1c079174b04de85e4153a6795bbe876ad9b0 for WebGL is specific to WebGL,
making the attachment work as stencil & depth attachment. However a
combination of both is not supported by pure GL ES 2.0 it seems.

However it appears that the easiest solution is to simply drop the depth
buffer. That appears to work well with desktop GL, a GL ES 2.0 context
(with mesa at least) and WebGL 1 (in Safari). Since femtovg doesn't use
depth testing, that seems like a reasonable compromise.