Fix rendering of unclipped images when they are scaled and cropped
The case a sub-rect of an image that begins at (0,0) is blitted
unclipped with a scale set on the canvas isn't handled correctly yet, as
the (u, v) texture coordinates don't take the scale into account yet. So
omit this for now from the optimization and deal with it in a later
commit.