Improve performance of the OpenGL renderer by splitting up the fragment shader
The fragment shader is large and riddled with branches. This doesn't
perform well on older GPUs and makes it harder to add more
specializations for faster rendering code paths. Already the image
filtering feels a bit out of place.

So this patch paves the way by compiling the fragment shader once for
each shader type, at the expensive of program changes at run-time.