Make it possible to construct gradient color stops more efficiently
When constructing the color stops from an iterator, one would previously
have to collect them into a vector, in order to be able to pass them as
a slice. While the implementation in Paint would collect them once more
into the Rc slice. By changing the API to take an iterator this can be
avoided.

In order to minimize the impact on existing code - such as the examples
- that pass slices to arrays, the API takes an IntoIterator. With the
2021 edition that allows passing the arrays directly.