Remove ouroboros dependency
With the two parent changes, it's now possible to hold a `&Font` and a
`rustybuzz::Face<'_>` that in turn references data from the `&Font`,
while populating the glyph cache at the same time, using interior
mutability. That way we can remove ouroboros and create the face on the
fly. For shaping it's crated once for all glyphs (faster), but for the
direct rendering we unfortunately have to create it once per glyph.
That's technically slower than the face that was cached in the `Font`
all time time, but it might be acceptable in exchange for a lighter
build and lower memory usage - especially in the light that direct
rendering is rather slow anyway (should be the exception).